Revert "Revert "Merge branch 'master' of https://gitlab.univ-nantes.fr/E177646T/googlehashcode2018""
This reverts commit ed509e130440c0e4d519e370ca3ca2faf9c41411.

+++ GoogleHashCode2018/Building.cpp
@@ -239,13 +239,13 @@ const vector<Coord>& Building::getCases() const
 float Building::getRatio()
 {
 	if(ratio<0)
-		ratio = nbCells / (rowNum * columnNum);
+		ratio = (float) (nbCells / (rowNum * columnNum));
 	return ratio;
 }
 float Building::getDensity()
 {
 	if(density<0)
-		density = nbCells/extra;
+		density = (float) (nbCells / extra);
 	return density;
 }
 unsigned int Building::getNbCells() const
